quiz_service/requirements.txt
flask
flask-cors
python-dotenv
pymongo
requests
redis